lrc files

+++ Nekko0/polls/views.py
@@ -432,7 +432,7 @@ def AddOrReplyMsg(request):
     # return dict
     # 包含基本的list 和 commentId && replyId
     userInfo = Userinfo.objects.get(id=userId)
-    infoList = [userInfo.username, userInfo.sex, userInfo.com_power/COM_POWER_RATE, userInfo.level_tag, \
+    infoList = [userInfo.username, userInfo.sex, int(userInfo.com_power/COM_POWER_RATE), userInfo.level_tag, \
                 userInfo.avatar_url]
     ret_json = {'userId': userId, 'date': timeNow.split('.')[0], \
                 'username': infoList[0], 'sex': infoList[1], 'power': infoList[2], \
@@ -532,7 +532,7 @@ class MsgBoardListView(ListView, FormView):
                 userInfo = Userinfo.objects.get(id=userid)
                 # 加入username, sex, level, level_tag, avatar_url
                 # com_power是算力值，需要换算成level
-                append_list = [userInfo.username, userInfo.sex, userInfo.com_power/COM_POWER_RATE, \
+                append_list = [userInfo.username, userInfo.sex, int(userInfo.com_power/COM_POWER_RATE), \
                                userInfo.level_tag, userInfo.avatar_url]
                 # append_list = [userInfo.username, userInfo.sex, userInfo.com_power, \
                 #                userInfo.level_tag, userInfo.avatar_url]
@@ -680,7 +680,7 @@ class CodeDetailView(ListView, FormView):
                 userInfo = Userinfo.objects.get(id=userid)
                 # 加入username, sex, level, level_tag, avatar_url
                 # com_power是算力值，需要换算成level
-                append_list = [userInfo.username, userInfo.sex, userInfo.com_power/COM_POWER_RATE, \
+                append_list = [userInfo.username, userInfo.sex, int(userInfo.com_power/COM_POWER_RATE), \
                                userInfo.level_tag, userInfo.avatar_url]
                 pickle_reply_list_item += append_list
                 # 判断子评论的所属者是否是当前用户
@@ -771,7 +771,7 @@ class DonateView(ListView, FormView):
                 userInfo = Userinfo.objects.get(id=userid)
                 # 加入username, sex, level, level_tag, avatar_url
                 # com_power是算力值，需要换算成level
-                append_list = [userInfo.username, userInfo.sex, userInfo.com_power/COM_POWER_RATE, \
+                append_list = [userInfo.username, userInfo.sex, int(userInfo.com_power/COM_POWER_RATE), \
                                userInfo.level_tag, userInfo.avatar_url]
                 pickle_reply_list_item += append_list
                 # 判断子评论的所属者是否是当前用户
@@ -900,7 +900,7 @@ class ArticleDetailView(ListView, FormView):
                 userInfo = Userinfo.objects.get(id=userid)
                 # 加入username, sex, level, level_tag, avatar_url
                 # com_power是算力值，需要换算成level
-                append_list = [userInfo.username, userInfo.sex, userInfo.com_power/COM_POWER_RATE, \
+                append_list = [userInfo.username, userInfo.sex, int(userInfo.com_power/COM_POWER_RATE), \
                                userInfo.level_tag, userInfo.avatar_url]
                 pickle_reply_list_item += append_list
                 # 判断子评论的所属者是否是当前用户
